requirements.txt
blinker==1.6.2
click==8.1.3
Deprecated==1.2.14
Flask==2.3.2
importlib-metadata==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
opentelemetry-api==1.18.0
opentelemetry-distro==0.39b0
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-aws-lambda==0.39b0
opentelemetry-instrumentation-dbapi==0.39b0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-instrumentation-jinja2==0.39b0
opentelemetry-instrumentation-logging==0.39b0
opentelemetry-instrumentation-sqlite3==0.39b0
opentelemetry-instrumentation-urllib==0.39b0
opentelemetry-instrumentation-wsgi==0.39b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
typing-extensions==4.6.3
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.15.0
